rename metrics to analytics, pt 1
